Remove per-service config/init casing from config manager and implement on the model class.
Implement dynamic handlers as PM instances if command style is gravity.

Drop the multi-instance "instance name" I implemented last night because
it doesn't really make sense for the dynamic handlers case. If people
really don't like automatic instance naming as a number from 0 we can
always give them the option to disable PM instances entirely (as is the
case with command style = direct, but we can support it with command
style = gravity as well if necessary).